requirements.txt
fastapi
uvicorn
python-multipart
pydantic
supabase
python-dotenv
python-jose[cryptography]
